refactor: restructure user authentication UI and rename OptionsContainerModule to OptionsContainerWidge
