--- lib/src/views/more/widgets/logout_btn_widget.dart
@@ -1,24 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-import '../../../providers/auth/auth_provider.dart';
-
-class LogoutButtonWidget extends ConsumerWidget {
-  const LogoutButtonWidget({super.key});
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    return Column(
-      children: [
-        ListTile(
-          dense: true,
-          contentPadding: EdgeInsets.zero,
-          leading: const Icon(Icons.logout),
-          title: const Text('로그아웃'),
-          onTap: () async {
-            ref.read(authProvider.notifier).logout();
-          },
-        ),
-      ],
-    );
-  }
-}
