+++ lib/src/views/more/widgets/user_auth_widget.dart
@@ -2,6 +2,8 @@ import 'package:flutter/material.dart';
 
 import '../../../../core/enum/mxnRate.dart';
 import '../../../../core/widgets/mxnContainer.dart';
+import 'app_chor/privacy_policy_widget.dart';
+import 'user_auth/logout_btn_widget.dart';
 
 class UserAuthWidget extends StatelessWidget {
   const UserAuthWidget({super.key});
@@ -14,13 +16,10 @@ class UserAuthWidget extends StatelessWidget {
         color: Colors.white,
         padding: EdgeInsets.symmetric(vertical: 8, horizontal: 16),
         child: Column(
-          // mainAxisAlignment: MainAxisAlignment.spaceEvenly,
           children: [
-            //LogoutButtonWidget(),
+            PrivacyPolicyWidget(),
             Divider(),
-            // LogoutButton(),
-            // Divider(),
-            // LogoutButton(),
+            LogoutButtonWidget(),
           ],
         ),
       ),
